Clarification on database schema documentation
Almost everything here looks good to me, but there's some ambiguity in the wording near the start that I have changed. If my change is correct, then incorporating it should improve the clarity of the docs, and if not, we need to be clearer about how versioning of surveys is done.